requirements.txt
asarPy==1.0.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
h5py==3.13.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
perlin-noise==1.13
pillow==11.0.0
pyparsing==3.2.2
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.2.0
six==1.17.0
sympy==1.13.1
torch==2.6.0+rocm6.2.4
torchaudio==2.6.0+rocm6.2.4
torchvision==0.21.0+rocm6.2.4
typing_extensions==4.12.2
